[refactor] move InvalidStudentProps to module types
