+++ 1_The_Physical/Level_2_Stateful_Guess_Point/Shared/Student/src/firstName.spec.ts
@@ -1,4 +1,5 @@
-import { FirstName, InvalidFirstName, InvalidStudentProps } from "./firstName";
+import { FirstName, InvalidFirstName } from "./firstName";
+import { InvalidStudentProps } from "./types";
 
 describe("FirstName Value Object", () => {
   it("should be defined when a valid input is given", () => {
