+++ 1_The_Physical/Level_2_Stateful_Guess_Point/Shared/Student/src/student.ts
@@ -1,4 +1,5 @@
-import { FirstName, InvalidFirstName, InvalidStudentProps } from "./firstName";
+import { FirstName, InvalidFirstName } from "./firstName";
+import { InvalidStudentProps } from "./types";
 
 export interface StudentProps {
   firstName: string;
