+++ 1_The_Physical/Level_2_Stateful_Guess_Point/Shared/Student/src/types.ts
@@ -1,0 +1,4 @@
+export type InvalidStudentProps<T> = {
+  type: T;
+  message: string;
+};
